Revert "Revert "[TS migration] Migrate withFullTransactionOrNotFound HOC to TypeScript""
This reverts commit 7b284fff864d45e96a07ad4273d25051ca878cf0.
